# Add index to statuses_tags#status_id