--- a/db/migrate/20170424112722_add_status_id_index_to_statuses_tags.rb
+++ b/db/migrate/20170424112722_add_status_id_index_to_statuses_tags.rb
@@ -1,0 +1,5 @@
+class AddStatusIdIndexToStatusesTags < ActiveRecord::Migration[5.0]
+  def change
+    add_index :statuses_tags, :status_id
+  end
+end