Optimized statuses_tags.

```
# pawoo.net
./bin/rails c production
$ class StatusesTag < ApplicationRecord; end
$ StatusesTag.count 
#=> 276146
```

```
EXPLAIN ANALYZE SELECT "statuses_tags".* FROM "statuses_tags"."status_id" IN (938537, 576666, 295882, 221478);

**before**
 Seq Scan on statuses_tags  (cost=0.00..5826.19 rows=11 width=12) (actual time=30.804..30.804 rows=0 loops=1)
   Filter: (status_id = ANY ('{938537,576666,295882,221478}'::bigint[]))
   Rows Removed by Filter: 276146
 Planning time: 110.067 ms
 Execution time: 30.880 ms
(5 rows)


**after**
 Index Scan using index_statuses_tags_on_status_id on statuses_tags  (cost=0.42..21.91 rows=11 width=12) (actual time=0.053..0.053 rows=0 loops=1)
   Index Cond: (status_id = ANY ('{938537,576666,295882,221478}'::bigint[]))
 Planning time: 0.199 ms
 Execution time: 0.087 ms
(4 rows)
```